general refactor

+++ tools/transformer_judo/staf_eval_allframes.py
@@ -15,117 +15,15 @@ from sklearn.metrics import confusion_matrix, classification_report
 import torch
 import torch.nn as nn
 from torch.utils.tensorboard import SummaryWriter
-import cv2
-from PIL import Image
 
 sys.path.append(os.getcwd())
 from lib import utils as utl
 from configs.judo import parse_model_args as parse_args
-from lib.utils.visualize import plot_bar, plot_to_image, add_pr_curve_tensorboard, get_segments
+from lib.utils.visualize import plot_bar, plot_to_image, add_pr_curve_tensorboard, get_segments, show_video_predictions
 from lib.models import build_model
 
 def to_device(x, device):
     return x.unsqueeze(0).to(device)
-
-def save_video_predictions(args,
-                           video_name,
-                           target_metrics,
-                           score_metrics,
-                           attn_weights,
-                           frames_dir='video_frames_25fps',
-                           fps=25):
-    pred_metrics = torch.argmax(torch.tensor(score_metrics), dim=1)
-    target_metrics = torch.argmax(torch.tensor(target_metrics), dim=1)
-
-    print('Loading and saving video: ' + video_name + ' . . . .\n.\n.')
-    frames = []
-    num_frames = target_metrics.shape[0]
-    idx = 0
-    count_chunk = 0
-    count_steps = 0
-    while idx < num_frames:
-        idx_frame = idx * args.chunk_size + args.chunk_size // 2
-        pil_frame = Image.open(osp.join(args.data_root,
-                                        frames_dir,
-                                        video_name,
-                                        str(idx_frame + 1) + '.jpg')).convert('RGB')
-
-        open_cv_frame = np.array(pil_frame)
-
-        # Convert RGB to BGR
-        open_cv_frame = open_cv_frame[:, :, ::-1].copy()
-
-        H, W, _ = open_cv_frame.shape
-        open_cv_frame = cv2.resize(open_cv_frame, (W // 2, H // 2), interpolation=cv2.INTER_AREA)
-
-        open_cv_frame = cv2.copyMakeBorder(open_cv_frame, 60, 0, 30, 30, borderType=cv2.BORDER_CONSTANT, value=0)
-        pred_label = args.class_index[pred_metrics[idx]]
-        target_label = args.class_index[target_metrics[idx]]
-
-        cv2.putText(open_cv_frame,
-                    pred_label,
-                    (0, 20),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.8,
-                    (0, 255, 0) if pred_label == target_label else (0, 0, 255),
-                    1)
-        cv2.putText(open_cv_frame,
-                    target_label,
-                    (0, 50),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.8,
-                    (255, 255, 255),
-                    1)
-        cv2.putText(open_cv_frame,
-                    'prob:{:.2f}'.format(torch.tensor(score_metrics)[idx, pred_metrics[idx]].item()),
-                    (210, 20),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.5,
-                    (0, 255, 0) if pred_label == target_label else (0, 0, 255),
-                    1)
-
-        cv2.putText(open_cv_frame,
-                    'weight:{:.2f}'.format(attn_weights[count_chunk].item()),
-                    (open_cv_frame.shape[1]-300, 20),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.8,
-                    (0, 0, 255),
-                    1)
-        cv2.putText(open_cv_frame,
-                    str(count_chunk%args.steps)+'/'+str(args.steps-1),
-                    (open_cv_frame.shape[1]-300, 50),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.8,
-                    (0, 0, 255),
-                    1)
-
-        # [ (idx_frame + 1) / 25 ]    => 25 because frames has been extracted at 25 fps
-        cv2.putText(open_cv_frame,
-                    '{:.2f}s'.format((idx_frame + 1) / fps),
-                    (295, 40),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.3,
-                    (255, 255, 255),
-                    1)
-        cv2.putText(open_cv_frame,
-                    str(idx_frame + 1),
-                    (295, 50),
-                    cv2.FONT_HERSHEY_SIMPLEX,
-                    0.3,
-                    (255, 255, 255),
-                    1)
-
-        frames.append(open_cv_frame)
-        if idx % 9 == 9-1:
-            count_chunk += 1
-        idx += 1
-
-    H, W, _ = open_cv_frame.shape
-    out = cv2.VideoWriter(video_name, cv2.VideoWriter_fourcc(*'mp4v'), fps / args.chunk_size, (W, H))
-    for frame in frames:
-        out.write(frame)
-    out.release()
-    print('. . . video saved at ' + os.path.join(os.getcwd(), video_name))
 
 def main(args):
     os.environ['CUDA_VISIBLE_DEVICES'] = args.gpu
@@ -142,10 +40,10 @@ def main(args):
     model.load_state_dict(checkpoint['model_state_dict'])
     model.train(False)
 
-    if not args.save_video:
+    if not args.show_predictions and not args.save_video:
         # i.e. we want to do a full evaluation of the test set and then compute stuff like confusion matrix, etc..
         tensorboard_dir = args.checkpoint.split('/')[:-1]
-        eval_dir = osp.join(*tensorboard_dir, 'eval_allframes')
+        eval_dir = osp.join(*tensorboard_dir, 'staf_eval_allframes')
         if osp.isdir(eval_dir):
             shutil.rmtree(eval_dir)
         os.mkdir(eval_dir)
@@ -156,17 +54,6 @@ def main(args):
 
     softmax = nn.Softmax(dim=1).to(device)
 
-    if args.eval_on_untrimmed:
-        args.test_session_set = args.test_session_set['UNTRIMMED']
-    else:
-        if args.use_untrimmed == args.use_trimmed == True:
-            args.test_session_set['TRIMMED'].extend(args.test_session_set['UNTRIMMED'])
-            args.test_session_set = args.test_session_set['TRIMMED']
-        elif args.use_trimmed:
-            args.test_session_set = args.test_session_set['TRIMMED']
-        elif args.use_untrimmed:
-            args.test_session_set = args.test_session_set['UNTRIMMED']
-
     utl.set_seed(int(args.seed))
     random.shuffle(args.test_session_set)
 
@@ -175,22 +62,14 @@ def main(args):
     if args.save_video:
         # when this option is activated, we evaluate and save ONLY ONE video, without showing it.
         args.test_session_set = args.test_session_set[:1]
+        args.show_predictions = False
+    if args.show_predictions:
+        count_frames = 0
 
     for session_idx, session in enumerate(args.test_session_set, start=1):
-        if 'Grand Prix' in session:
-            dataset_type = 'UNTRIMMED'
-        elif 'GoPro' in session:
-            dataset_type = 'TRIMMED'
-        else:
-            raise Exception('Unknown video name: ' + session)
-        if not osp.isfile(osp.join(args.data_root, dataset_type, args.model_target, session + '.npy')):
-            # skip videos in which the pose model does not detect any fall(i.e. fall==-1  in fall_detections.csv).
-            # TODO: fix these videos later on, in order to incluso also them
-            continue
-
         start = time.time()
         with torch.set_grad_enabled(False):
-            original_target = np.load(osp.join(args.data_root, dataset_type, args.model_target, session + '.npy'))
+            original_target = np.load(osp.join(args.data_root, args.model_target, session + '.npy'))
             # round to multiple of CHUNK_SIZE
             num_frames = original_target.shape[0]
             num_frames = num_frames - (num_frames % args.chunk_size)
@@ -198,38 +77,69 @@ def main(args):
             # For each chunk, take only the central frame
             target = original_target[args.chunk_size // 2::args.chunk_size]
 
-            features_extracted = np.load(osp.join(args.data_root, dataset_type, args.model_input, session + '.npy'),
+            # it contains staf predictions, i.e. background or candidate
+            candidate_actions = np.load(osp.join(args.data_root, 'candidatesV2ALL_' + args.model_target, session + '.npy'))
+            # round to multiple of CHUNK_SIZE
+            num_frames = candidate_actions.shape[0]
+            num_frames = num_frames - (num_frames % args.chunk_size)
+            candidate_actions = candidate_actions[:num_frames]
+            # For each chunk, take only the central frame
+            candidate_actions = candidate_actions[args.chunk_size // 2::args.chunk_size]
+
+            features_extracted = np.load(osp.join(args.data_root, args.model_input, session + '.npy'),
                                          mmap_mode='r')
             features_extracted = torch.as_tensor(features_extracted.astype(np.float32))
+            if 'resnext' in args.model_input:
+                num_frames = feature_vectors.shape[0]
+                num_frames = num_frames - (num_frames % args.chunk_size)
+                feature_vectors = feature_vectors[:num_frames]
+                feature_vectors = feature_vectors[args.chunk_size // 2::args.chunk_size]
 
+            COOLDOWN = 0
             samples = []
-            attn_weights = []
             for count in range(target.shape[0]):
-                samples.append(features_extracted[count])
 
-                if count % args.steps == 0 and count != 0:
-                    samples = torch.stack(samples).unsqueeze(0).to(device)
-                    scores, attn_weights_t = model(samples)         # scores.shape == (1, steps, num_classes)
-                                                                    # attn_weights_t.shape == (batch_size, steps)
-                    attn_weights.append(attn_weights_t)
+                if candidate_actions[count, 1] == 1 and COOLDOWN == 0:
+                    # first step
+                    samples.append(features_extracted[count])
+                    COOLDOWN = args.steps - 1
+                else:
+                    if COOLDOWN > 0:
+                        # subsequent steps
+                        COOLDOWN -= 1
+                        samples.append(features_extracted[count])
+                        if len(samples) == args.steps:
+                            samples = torch.stack(samples)
+                            samples = to_device(samples, device)    # samples.shape == (1, steps, feat_vect_dim)
+                            scores = model(samples)
 
-                    scores = scores.squeeze(0)
-                    scores = softmax(scores).cpu().detach().numpy()
-                    for s in range(len(scores)):
-                        for c in range(args.chunk_size):
-                            score_metrics.append(scores[s])
-                    samples = []
+                            scores = scores.squeeze(0)
+                            scores = softmax(scores).cpu().detach().numpy()
+                            for s in range(len(scores)):
+                                for c in range(args.chunk_size):
+                                    score_metrics.append(scores[s])
+
+                            samples = []
+                    else:
+                        # get staf prediction, i.e. background
+                        scores = torch.zeros(args.steps, args.num_classes, dtype=torch.float32)
+                        scores[:, 0] = 100
+                        scores = softmax(scores).numpy()
+                        for s in range(len(scores)):
+                            for c in range(args.chunk_size):
+                                score_metrics.append(scores[s])
 
                 for c in range(args.chunk_size):
                     target_metrics.append(original_target[count * args.chunk_size + c])
 
-            if samples != []:
+            if len(samples) != 0:
                 appo = len(samples)
                 for i in range(appo, args.steps):
                     samples.append(torch.zeros_like(features_extracted[0]))
-                samples = torch.stack(samples).unsqueeze(0).to(device)
-                scores, attn_weights_t = model(samples)
-                attn_weights.append(attn_weights_t[:, :appo])
+                samples = torch.stack(samples)
+                samples = to_device(samples, device)
+                scores = model(samples)
+
                 scores = scores[:, :appo]
                 scores = scores.squeeze(0)
                 scores = softmax(scores).cpu().detach().numpy()
@@ -243,20 +153,27 @@ def main(args):
                                                                                  session_idx,
                                                                                  len(args.test_session_set),
                                                                                  end - start))
+        if args.show_predictions:
+            appo = args.chunk_size
+            args.chunk_size = 1
+            show_video_predictions(args,
+                                   session,
+                                   target_metrics[count_frames:count_frames + original_target.shape[0]],
+                                   score_metrics[count_frames:count_frames + original_target.shape[0]],
+                                   frames_dir='video_frames_25fps',
+                                   fps=25)
+            args.chunk_size = appo
+            count_frames += original_target.shape[0]
+
     if args.save_video:
         # here the video will be saved
-        attn_weights = torch.cat(attn_weights, dim=1).squeeze(0)
         args.chunk_size = 1
-        appo = args.data_root
-        args.data_root = args.data_root + '/' + dataset_type
-        save_video_predictions(args,
+        show_video_predictions(args,
                                session,
                                target_metrics,
                                score_metrics,
-                               attn_weights=attn_weights,
                                frames_dir='video_frames_25fps',
                                fps=25)
-        args.data_root = appo
         # print some stats about the video labels and predictions, then kill the program
         print('\n=== LABEL SEGMENTS ===')
         segments_list = get_segments(target_metrics, args.class_index, 25, args.chunk_size)
